Fix Oscillator sync

+++ plugins/organic/organic.h
@@ -154,8 +154,8 @@ private:
 		MM_OPERATORS
 		Oscillator * oscLeft;
 		Oscillator * oscRight;
-		float phaseOffsetLeft[NUM_OSCILLATORS];
-		float phaseOffsetRight[NUM_OSCILLATORS];		
+		double phaseOffsetLeft[NUM_OSCILLATORS];
+		double phaseOffsetRight[NUM_OSCILLATORS];
 	} ;
 
 	const IntModel m_modulationAlgo;
